40 create mock data script for the db

+++ database/tables/files.sql
@@ -1,8 +1,8 @@
 CREATE TABLE files (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     path TEXT NOT NULL UNIQUE,
     basename TEXT NOT NULL,
     directory TEXT NOT NULL,
-    size BIGINT NOT NULL,
+    size INTEGER NOT NULL,
     mtime TIMESTAMPTZ NOT NULL
 );
